+++ src/realm/util/file.cpp
@@ -671,7 +671,7 @@ size_t File::read(char* data, size_t size)
         uint64_t pos_original = File::get_file_pos(m_fd);
         REALM_ASSERT(!int_cast_has_overflow<size_t>(pos_original));
         size_t pos = size_t(pos_original);
-        Map<char> read_map(*this, access_ReadOnly, static_cast<size_t>(pos + size));
+        Map<char> read_map(*this, access_ReadOnly, (pos + size));
         realm::util::encryption_read_barrier(read_map, pos, size);
         memcpy(data, read_map.get_addr() + pos, size);
         uint64_t cur = File::get_file_pos(m_fd);
@@ -740,7 +740,7 @@ void File::write(const char* data, size_t size)
         uint64_t pos_original = get_file_pos(m_fd);
         REALM_ASSERT(!int_cast_has_overflow<size_t>(pos_original));
         size_t pos = size_t(pos_original);
-        Map<char> write_map(*this, access_ReadWrite, static_cast<size_t>(pos + size));
+        Map<char> write_map(*this, access_ReadWrite, (pos + size));
         realm::util::encryption_read_barrier(write_map, pos, size);
         memcpy(write_map.get_addr() + pos, data, size);
         realm::util::encryption_write_barrier(write_map, pos, size);
